+++ res/shaders/waterFragment.glsl
@@ -2,6 +2,8 @@
 
 in vec4 clipSpace;
 in vec2 textureCoords;
+in vec3 toCameraVector;
+in vec3 fromLightVector;
 
 out vec4 out_Color;
 
@@ -9,20 +11,35 @@ uniform sampler2D reflectionTexture;
 uniform sampler2D refractionTexture;
 
 uniform sampler2D dudvMap;
-uniform float moveFactorOffset;
+uniform sampler2D normalMap;
+uniform sampler2D depthMap;
 
-const float waveStrength = 0.02;
+uniform float moveFactorOffset;
+uniform float near;
+uniform float far;
+
+uniform vec3 lightColor;
+
+const float waveStrength = 0.04;
+const float shineDamper = 20.0;
+const float reflectivity = 0.5;
 
 void main(void){
-    vec2 normDeviceSpace = clipSpace.xy / clipSpace.w;
-    vec2 screenSpaceCoords = normDeviceSpace / 2.0 + 0.5;
+    vec2 screenSpaceCoords = (clipSpace.xy / clipSpace.w) / 2.0 + 0.5;
 
     vec2 refractTexCoords = vec2(screenSpaceCoords.xy);
     vec2 reflectTexCoords = vec2(screenSpaceCoords.x, -screenSpaceCoords.y);
 
-    vec2 distortion1 = (texture(dudvMap, vec2(textureCoords.x + moveFactorOffset, textureCoords.y)).rg * 2.0 - 1.0) * waveStrength;
-    vec2 distortion2 = (texture(dudvMap, vec2(-textureCoords.x + moveFactorOffset, textureCoords.y + moveFactorOffset)).rg * 2.0 - 1.0) * waveStrength;
-    vec2 totalDistortion = distortion1 + distortion2;
+    float depth = texture(depthMap, refractTexCoords).r;
+    float floorDistance = 2.0 * near * far / (far + near - (2.0 * depth - 1.0) * (far - near));
+    
+    depth = gl_FragCoord.z;
+    float waterDistance = 2.0 * near * far / (far + near - (2.0 * depth - 1.0) * (far - near));
+    float waterDepth = floorDistance - waterDistance;
+
+    vec2 distortedTexCoords = texture(dudvMap, vec2(textureCoords.x + moveFactorOffset, textureCoords.y)).rg * 0.1;
+    distortedTexCoords = textureCoords + vec2(distortedTexCoords.x, distortedTexCoords.y + moveFactorOffset);
+    vec2 totalDistortion = (texture(dudvMap, distortedTexCoords).rg * 2.0 - 1.0) * waveStrength * clamp(waterDepth / 20.0, 0.0, 1.0);
 
     refractTexCoords += totalDistortion;
     refractTexCoords = clamp(refractTexCoords, 0.001, 0.999);
@@ -34,6 +51,21 @@ void main(void){
 
     vec4 reflectColor = texture(reflectionTexture, reflectTexCoords);
     vec4 refractColor = texture(refractionTexture, refractTexCoords);
-    out_Color = mix(reflectColor, refractColor, 0.5);
-    out_Color = mix(out_Color, vec4(0.0, 0.3, 0.5, 1.0), 0.2);
+
+    vec4 normalMapColor = texture(normalMap, distortedTexCoords);
+    vec3 normal = vec3(normalMapColor.r * 2.0 - 1.0, normalMapColor.b * 3.0, normalMapColor.g * 2.0 - 1.0);
+    normal = normalize(normal);
+
+    vec3 viewVector = normalize(toCameraVector);
+    float refractiveFactor = dot(viewVector, normal);
+    refractiveFactor = clamp(pow(refractiveFactor, 5.0), 0.0, 1.0);
+
+    vec3 reflectedLight = reflect(normalize(fromLightVector), normal);
+    float specular = max(dot(reflectedLight, viewVector), 0.0);
+    specular = pow(specular, shineDamper);
+    vec3 specularHighlights = lightColor * specular * reflectivity * clamp(waterDepth / 5.0, 0.0, 1.0);
+
+    out_Color = mix(reflectColor, refractColor, refractiveFactor);
+    out_Color = mix(out_Color, vec4(0.0, 0.3, 0.5, 1.0), 0.2) + vec4(specularHighlights, 0.0);
+    out_Color.a = clamp(waterDepth / 5.0, 0.0, 1.0);
 }