+++ src/hr/avrbanac/openglplayground/shaders/WaterShader.java
@@ -1,15 +1,15 @@
 
 package hr.avrbanac.openglplayground.shaders;
 
-import static hr.avrbanac.openglplayground.Globals.WATER_FRAGMENT_FILE;
-import static hr.avrbanac.openglplayground.Globals.WATER_VERTEX_FILE;
+import static hr.avrbanac.openglplayground.Globals.*;
 import hr.avrbanac.openglplayground.entities.Camera;
+import hr.avrbanac.openglplayground.entities.Light;
 import hr.avrbanac.openglplayground.maths.Matrix4f;
 
 /**
  *
  * @author avrbanac
- * @version 1.0.16
+ * @version 1.0.17
  */
 public class WaterShader extends ShaderProgram {
     private int locationModelMatrix;
@@ -19,7 +19,14 @@ public class WaterShader extends ShaderProgram {
     private int locationRefractionTexutre;
     private int locationDuDvMap;
     private int locationMoveFactor;
- 
+    private int locationCameraPosition;
+    private int locationNormalMap;
+    private int locationLightPosition;
+    private int locationLightColor;
+    private int locationDepthMap;
+    private int locationNear;
+    private int locationFar;
+    
     public WaterShader() {
         super(WATER_VERTEX_FILE, WATER_FRAGMENT_FILE);
     }
@@ -38,6 +45,18 @@ public class WaterShader extends ShaderProgram {
         locationRefractionTexutre   = getUniformLocation("refractionTexture");
         locationDuDvMap             = getUniformLocation("dudvMap");
         locationMoveFactor          = getUniformLocation("moveFactorOffset");
+        locationCameraPosition      = getUniformLocation("cameraPosition");
+        locationNormalMap           = getUniformLocation("normalMap");
+        locationLightPosition       = getUniformLocation("lightPosition");
+        locationLightColor          = getUniformLocation("lightColor");
+        locationDepthMap            = getUniformLocation("depthMap");
+        locationNear                = getUniformLocation("near");
+        locationFar                 = getUniformLocation("far");
+    }
+    
+    public void loadNearFarPlane() {
+        super.loadFloat(locationNear, NEAR_PLANE);
+        super.loadFloat(locationFar, FAR_PLANE);
     }
     
     public void loadMoveFactorOffset(float factor) {
@@ -48,6 +67,8 @@ public class WaterShader extends ShaderProgram {
         super.loadInt(locationReflectionTexutre, 0);
         super.loadInt(locationRefractionTexutre, 1);
         super.loadInt(locationDuDvMap, 2);
+        super.loadInt(locationNormalMap, 3);
+        super.loadInt(locationDepthMap, 4);
     }
  
     public void loadProjectionMatrix(Matrix4f projection) {
@@ -57,10 +78,16 @@ public class WaterShader extends ShaderProgram {
     public void loadViewMatrix(Camera camera){
         Matrix4f viewMatrix = Matrix4f.view(camera);
         loadMatrix(locationViewMatrix, viewMatrix);
+        super.load3DVector(locationCameraPosition, camera.getPosition());
     }
  
     public void loadModelMatrix(Matrix4f modelMatrix){
         loadMatrix(locationModelMatrix, modelMatrix);
+    }
+    
+    public void loadLight(Light sun) {
+        super.load3DVector(locationLightPosition, sun.getPosition());
+        super.load3DVector(locationLightColor, sun.getColor());
     }
  
 }